refactor(cmd): move commands to internal package
Both interactive and non-interactive commands now
lives as separate structs inside internal/cmd.

This change reduces the complexity of the cobra
entry point and pave the way to introduce tests.

Additional changes:
renamed the gci prefix for internal imports from
emilkje/cwc to intility/cwc to properly format the
imports.
