+++ cmd/config.go
@@ -6,10 +6,11 @@ import (
 	"strconv"
 	"strings"
 
+	"github.com/spf13/cobra"
+
 	"github.com/intility/cwc/pkg/config"
 	"github.com/intility/cwc/pkg/errors"
 	"github.com/intility/cwc/pkg/ui"
-	"github.com/spf13/cobra"
 )
 
 func createConfigCommand() *cobra.Command {
