+++ cmd/logout.go
@@ -3,8 +3,9 @@ package cmd
 import (
 	"fmt"
 
-	"github.com/intility/cwc/pkg/config"
 	"github.com/spf13/cobra"
+
+	"github.com/intility/cwc/pkg/config"
 )
 
 func createLogoutCmd() *cobra.Command {
